+++ components/collector/src/quality_time_collector.py
@@ -2,6 +2,8 @@
 
 import asyncio
 import logging
+import os
+from typing import NoReturn
 
 # Make sure subclasses are registered
 import metric_collectors  # pylint: disable=unused-import # lgtm [py/unused-import]
@@ -9,11 +11,12 @@ import source_collectors  # pylint: disable=unused-import # lgtm [py/unused-impo
 from base_collectors import Collector
 
 
-async def collect(log_level: int | None = None) -> None:
+async def collect() -> NoReturn:
     """Collect the measurements indefinitely."""
-    logging.getLogger().setLevel(logging.ERROR if log_level is None else log_level)
+    log_level = str(os.getenv("COLLECTOR_LOG_LEVEL", "WARNING"))  # cast to str to keep mypy happy
+    logging.getLogger().setLevel(log_level)
     await Collector().start()
 
 
 if __name__ == "__main__":
-    asyncio.run(collect(logging.WARNING))  # pragma: no cover
+    asyncio.run(collect())  # pragma: no cover